Fix test suit fail issue
