+++ app/frontend/src/components/designer/FormSettings.vue
@@ -476,7 +476,7 @@
 import { mapActions } from 'vuex';
 import { mapFields } from 'vuex-map-fields';
 import { IdentityMode, IdentityProviders, Regex } from '@/utils/constants';
-import { getAvailableDates } from '@/utils/permissionUtils';
+import { getAvailableDates } from '@/utils/transformUtils';
 import moment from 'moment';
 
 export default {
